pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skill-generator"
version = "0.1.0"
description = "AI CLI that generates Agent Skills for Codex and Claude Code using the OpenAI Agents SDK."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
  "openai-agents>=0.1.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.2",
  "rich>=13.7.1",
  "typer>=0.12.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[project.scripts]
skill-generator = "skillforge.cli:app"
skillforge = "skillforge.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/skillforge"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
